+++ CHANGELOG.md
@@ -6,10 +6,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.0] - 2018-12-26
+### Added
+- Addition of an optional live system test of a configurable job adapter
+
+### Fixed
+- Fix Torque adapter crash by fixing scope resolution on Attrl and Attropl
+- Fix SGE adapter crash in `OodCore::Job::Adapters::Sge::Batch#get_info_enqueued_job` when libdrmma is not available (DRMMA constant not defined)
+
+### Changed
+- Always set `SGE_ROOT` env var, for both SGE commands via popen and when using libdrmaa
+- Use libdrmaa only when libdrmaa is set in the cluster config
+
+
+## [0.6.0] - 2018-12-19
 ### Added
 - Added ability to override the default password length
-- Add notion of an aggregate status to OodCore::Job::Info to support job array parents
-- Changed order of OodCore::Job:Status.states to support precedence / sorting
+- Merge the pbs-ruby gem removing that as a dependency, but adding FFI
+- Added support for overriding resource manager client executables using `bin_overrides` in the cluster configs
+- Add support for the Grid Engine resource manager (tested on GE 6.2u5 and UGE 8.0.1)
 
 ### Fixed
 - Fixed a bug in password creation where certain locales resulted in invalid passwords [#91](https://github.com/OSC/ood_core/issues/91)
@@ -137,7 +153,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release!
 
-[Unreleased]: https://github.com/OSC/ood_core/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/OSC/ood_core/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/OSC/ood_core/compare/v0.6.0...v0.7.0
+[0.6.0]: https://github.com/OSC/ood_core/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/OSC/ood_core/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/OSC/ood_core/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/OSC/ood_core/compare/v0.3.0...v0.4.0